feat(axomdbook): more aggressively invalidate themes
This adds a checkThemeCacheValidity function to book.js that saves+loads
the display name of the default theme. If it changes we invalidate the
actual cached theme selection (causing the default_theme to be selected),
preventing confusing situations where it looks like mdbook is broken on
theme change because it was remembering the selected theme from before.