feat(generate-css): use local `tailwindcss` binary
Currently the `oranda-generate-css` crate will always attempt to
download a precompiled `tailwindcss` binary from GitHub Releases, and
store it in the user's cache directory. This doesn't work on platforms
such as NixOS, where the global `.cache` directory is not writable while
building a package. Thus, the Nix flake in the Oranda repo doesn't
actually work at all.

This commit changes the `oranda-generate-css` to first check if a binary
named `tailwindcss` is on the current `PATH`, and uses that rather than
downloading the precompiled binary if Tailwind is already available. We
check for the existence of `tailwindcss` by running `tailwindcss --help`
and testing if the exit code is 0, which seemed like the most
cross-platform way to check if the binary exists. Potentially, we may
want to change this to check for a specific `tailwindcss` version in the
future, but I noticed that the current code downloads the latest
release, rather than a fixed version, so I wanted the same behavior
here.

This will allow us to modify the Nix flake to depend on `tailwindcss`
from nixpkgs, so that the download is skipped. This should fix the flake
build.

I noticed that the tailwindlabs/tailwindcss repo also contains a [Rust
implementation of the Tailwind CLI][2]. We may, eventually, want to
consider changing the approach used here to simply depending on the Rust
tailwind implementation as a library dependency, so that it can be
resolved by Cargo instead of downloaded from GitHub. However, that's a
larger change, and I wanted to do the simplest change that would fix the
Nix flake. Additionally, it might be desirable to add some kind of CI
step that ensures the flake builds successfully, if the maintainers want
to officially support it. That way, we could avoid breaking the flake in
the future.

[1]: https://search.nixos.org/packages?channel=unstable&show=tailwindcss&from=0&size=50&sort=relevance&type=packages&query=tailwindcss
[2]: https://github.com/tailwindlabs/tailwindcss/blob/master/oxide/crates/cli/src/main.rs