fix(flake): depend on `tailwindcss` from nixpkgs
Follow up from 1e66acc. This commit changes the Nix flake to depend on
`tailwindcss` from nixpkgs instead of downloading a pre-built binary
from GitHub. Now that the build script will run a pre-existing
`tailwindcss` binary if one exists, this avoids the need to write to the
cache directory, which is not writable inside of a Nix flake, as well as
the precompiled binary, which may not work on NixOS at all.