fix: don't respect cargo-dist's install hints for curl-sh exprs
The dist-manifest.json fundamentally lacks the information to generate perfect pretty curl-sh exprs, because it needs to be uploaded before the create-release endpoint tells us this url format. As such, oranda should ignore the manifest for these install-hints, and instead use the inference backend for them, which can use the URLs returned by the list-releases endpoint (or the github release's base url).