fix: linkcheck plugin on *nix
On Linux and macOS, trying to run with the linkcheck plugin configured would
error out with the following error:

```
Error: Unable to resolve the source directory

Caused by:
    No such file or directory (os error 2)
  × Couldn't build your mdbook at book
  ├─▶ Rendering failed
  ╰─▶ The "linkcheck" renderer failed
  ```

This turns out to have been caused by us passing a relative path to the
book directory instead of an absolute path. The relative path was fine for
the main rendering process, but not the linkcheck plugin.